no rosout thread created if environment variable 'ROS_NO_ROSOUT' is exported.
polling frequency of PollManager is tunable via ROS_POLLMGR_POLL_TIMEOUT.

Abstract Named Socket support via env value named "ROS_UDS_EXT_FEATURE".

polling frequency of XMLRPCManager is tunable via ROS_XMLRPCMGR_POLL_TIMEOUT.

+++ clients/roscpp/src/libros/transport/transport_uds_datagram.cpp
@@ -72,12 +72,14 @@ TransportUDSDatagram::TransportUDSDatagram(PollSet* poll_set, int flags, int max
 
 TransportUDSDatagram::~TransportUDSDatagram()
 {
-  if (!server_uds_path_.empty())
-  {
-    int err = unlink(server_uds_path_.c_str());
-    if(err != 0)
+  if (!ROS_UDS_EXT_IS_ENABLE(ROS_UDS_EXT_ABSTRACT_SOCK_NAME)) {
+    if (!server_uds_path_.empty())
     {
+      int err = unlink(server_uds_path_.c_str());
+      if(err != 0)
+      {
         ROS_ERROR("Unable to remove %s: %s", server_uds_path_.c_str(), strerror(errno));
+      }
     }
   }
 
@@ -153,8 +155,14 @@ bool TransportUDSDatagram::connect(const std::string& uds_path, int connection_i
   memset(&sock_addr, 0, sizeof(sock_addr));
   sock_addr.sun_family = AF_UNIX;
 
-  strncpy(sock_addr.sun_path, uds_path.c_str(), uds_path.length());
-  int sock_length = offsetof(struct sockaddr_un, sun_path) + uds_path.length();
+  int sock_length;
+  if (ROS_UDS_EXT_IS_ENABLE(ROS_UDS_EXT_ABSTRACT_SOCK_NAME)) {
+    strncpy(&sock_addr.sun_path[1], uds_path.c_str(), uds_path.length());
+    sock_length = offsetof(struct sockaddr_un, sun_path) + uds_path.length() + 1;
+  } else {
+    strncpy(sock_addr.sun_path, uds_path.c_str(), uds_path.length());
+    sock_length = offsetof(struct sockaddr_un, sun_path) + uds_path.length();
+  }
 
   if (::connect(sock_, (sockaddr *)&sock_addr, sock_length))
   {
@@ -194,10 +202,17 @@ bool TransportUDSDatagram::createIncoming(bool is_server)
   memset(&sock_addr, 0, sizeof(sock_addr));
   sock_addr.sun_family = AF_UNIX;
 
-  static boost::atomic_uint32_t uds_counter(0);
-  server_uds_path_ = generateServerUDSPath(uds_counter++);
-  strncpy(sock_addr.sun_path, server_uds_path_.c_str(), server_uds_path_.length());
-  int sock_length = offsetof(struct sockaddr_un, sun_path) + server_uds_path_.length();
+  int sock_length;
+  if (ROS_UDS_EXT_IS_ENABLE(ROS_UDS_EXT_ABSTRACT_SOCK_NAME)) {
+    server_uds_path_ = generateServerUDSPath();
+    strncpy(&sock_addr.sun_path[1], server_uds_path_.c_str(), server_uds_path_.length());
+    sock_length = offsetof(struct sockaddr_un, sun_path) + server_uds_path_.length() + 1;
+  } else {
+    static boost::atomic_uint32_t uds_counter(0);
+    server_uds_path_ = generateServerUDSPath(uds_counter++);
+    strncpy(sock_addr.sun_path, server_uds_path_.c_str(), server_uds_path_.length());
+    sock_length = offsetof(struct sockaddr_un, sun_path) + server_uds_path_.length();
+  }
 
   if (bind(sock_, (sockaddr *)&sock_addr, sock_length) < 0)
   {
